pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "docs2llms"
version = "0.1.0"
description = "A tool to convert website documentation into LLM-friendly formats"
requires-python = ">=3.8"
dependencies = [
    "beautifulsoup4>=4.9.3",
    "httpx>=0.24.0",
    "scrapy>=2.8.0",
    "click>=8.0.0",
    "pyyaml>=6.0.0",
    "tqdm>=4.65.0",
    "markdown>=3.4.0",
    "html2text>=2020.1.16",
    "lxml>=4.9.0",
    "aiohttp>=3.8.0",
    "asyncio>=3.4.3",
    "ratelimit>=2.2.1",
    "cachetools>=5.0.0",
    "rich>=12.0.0",
    "loguru>=0.6.0",
    "playwright>=1.39.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.18.0",
    "pytest-cov>=3.0.0",
    "responses>=0.23.0",
    "black>=22.3.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
strict = true
ignore_missing_imports = true

[project.scripts]
docs2llms = "docs2llms.cli:main"